# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Commission Copilot API",
    "description": "Moteur de calcul des commissions de courtage et assistant conversationnel. Données fictives.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/perimeters/{code}": {
      "get": {
        "tags": [
          "paramétrage"
        ],
        "summary": "Get Perimeter",
        "operationId": "get_perimeter_api_perimeters__code__get",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PerimeterDetailsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/simulate": {
      "post": {
        "tags": [
          "calcul"
        ],
        "summary": "Simulate",
        "operationId": "simulate_api_simulate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalculationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/samples/{perimeter}/contracts/{contract_id}": {
      "get": {
        "tags": [
          "calcul"
        ],
        "summary": "Get Sample Contract",
        "operationId": "get_sample_contract_api_samples__perimeter__contracts__contract_id__get",
        "parameters": [
          {
            "name": "perimeter",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Perimeter"
            }
          },
          {
            "name": "contract_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Contract Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SampleContractOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chat": {
      "post": {
        "tags": [
          "assistant"
        ],
        "summary": "Chat",
        "operationId": "chat_api_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "429": {
            "description": "Limite de débit de la démo atteinte (en-tête Retry-After) ou quota du fournisseur de modèle épuisé"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chat/stream": {
      "post": {
        "tags": [
          "assistant"
        ],
        "summary": "Chat Stream",
        "operationId": "chat_stream_api_chat_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Flux Server-Sent Events. Événements, dans l'ordre où ils surviennent :\n\n- `delta` `{\"text\"}` : texte généré. Le texte qui précède des appels d'outils n'est pas la réponse finale.\n- `tool_call` `{\"name\", \"arguments\"}` : un outil va être exécuté.\n- `tool_result` : trace de l'appel (même format que `tool_calls` de `/api/chat`).\n- `done` : réponse complète, au même format que `/api/chat`.\n- `error` `{\"status\", \"detail\"}` : échec du modèle (429 quota épuisé, 502 indisponible) ; le flux s'arrête.\n",
            "content": {
              "text/event-stream": {}
            }
          },
          "429": {
            "description": "Limite de débit de la démo atteinte (en-tête Retry-After) ou quota du fournisseur de modèle épuisé"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CalculationOut": {
        "properties": {
          "perimeter": {
            "type": "string",
            "title": "Perimeter"
          },
          "month": {
            "type": "string",
            "title": "Month"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/CommissionLineOut"
            },
            "type": "array",
            "title": "Lines"
          },
          "exclusions": {
            "items": {
              "$ref": "#/components/schemas/ExclusionOut"
            },
            "type": "array",
            "title": "Exclusions"
          },
          "totals": {
            "$ref": "#/components/schemas/TotalsOut"
          }
        },
        "type": "object",
        "required": [
          "perimeter",
          "month",
          "lines",
          "exclusions",
          "totals"
        ],
        "title": "CalculationOut"
      },
...
      "ChatRequest": {
        "properties": {
          "messages": {
            "items": {
              "$ref": "#/components/schemas/ChatMessage"
            },
            "type": "array",
            "maxItems": 20,
            "minItems": 1,
            "title": "Messages"
          }
        },
        "type": "object",
        "required": [
          "messages"
        ],
        "title": "ChatRequest"
      },
      "ChatResponse": {
        "properties": {
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "tool_calls": {
            "items": {
              "$ref": "#/components/schemas/ToolCallTrace"
            },
            "type": "array",
            "title": "Tool Calls"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/CitationOut"
            },
            "type": "array",
            "title": "Citations",
            "description": "Règles citées dans la réponse ou appliquées par le moteur"
          },
          "unverified_amounts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Unverified Amounts",
            "description": "Montants cités par l'assistant qui ne proviennent d'aucune source (outil, règles, question)"
          },
          "unknown_rules": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Unknown Rules",
            "description": "Identifiants de règle cités qui n'existent pas"
          },
          "unknown_products": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Unknown Products",
            "description": "Noms de produits cités qui n'existent pas dans le catalogue"
          },
          "model": {
            "type": "string",
            "title": "Model"
          }
        },
        "type": "object",
        "required": [
          "answer",
          "tool_calls",
          "citations",
          "unverified_amounts",
          "unknown_rules",
          "unknown_products",
          "model"
        ],
        "title": "ChatResponse"
      },
...
      "CommissionLineOut": {
        "properties": {
          "contract_id": {
            "type": "string",
            "title": "Contract Id"
          },
          "product": {
            "type": "string",
            "title": "Product"
          },
          "commission_type": {
            "type": "string",
            "enum": [
              "P",
              "RP",
              "L"
            ],
            "title": "Commission Type"
          },
          "commission_label": {
            "type": "string",
            "title": "Commission Label"
          },
          "amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Amount"
          },
          "amount_display": {
            "type": "string",
            "title": "Amount Display"
          },
          "rule_id": {
            "type": "string",
            "title": "Rule Id"
          },
          "rule_label": {
            "type": "string",
            "title": "Rule Label"
          },
          "formula": {
            "type": "string",
            "title": "Formula"
          },
          "rate_source": {
            "type": "string",
            "title": "Rate Source"
          }
        },
        "type": "object",
        "required": [
          "contract_id",
          "product",
          "commission_type",
          "commission_label",
          "amount",
          "amount_display",
          "rule_id",
          "rule_label",
          "formula",
          "rate_source"
        ],
        "title": "CommissionLineOut"
      },
      "ContractInput": {
        "properties": {
          "product": {
            "type": "string",
            "maxLength": 60,
            "minLength": 1,
            "title": "Product",
            "description": "Code produit, ex. NS-CONFORT"
          },
          "state": {
            "type": "string",
            "maxLength": 40,
            "minLength": 1,
            "title": "State",
            "description": "AFN, SEF ou RES (libellés des assureurs acceptés)"
          },
          "subscription_date": {
            "type": "string",
            "format": "date",
            "title": "Subscription Date"
          },
          "effective_date": {
            "type": "string",
            "format": "date",
            "title": "Effective Date"
          },
          "annual_premium": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 10000000.0,
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Annual Premium",
            "description": "Prime annuelle hors taxes, en euros"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "guarantee": {
            "type": "string",
            "maxLength": 60,
            "title": "Guarantee",
            "default": ""
          },
          "segment": {
            "type": "string",
            "maxLength": 60,
            "title": "Segment",
            "default": ""
          },
          "file_rate_year_1": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 10.0,
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Rate Year 1",
            "description": "Taux transmis par l'assureur"
          },
          "contract_id": {
            "type": "string",
            "maxLength": 60,
            "minLength": 1,
            "title": "Contract Id",
            "default": "SIMULATION"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "product",
          "state",
          "subscription_date",
          "effective_date",
          "annual_premium"
        ],
        "title": "ContractInput"
      },
      "ContractRecordOut": {
        "properties": {
          "contract_id": {
            "type": "string",
            "title": "Contract Id"
          },
          "product": {
            "type": "string",
            "title": "Product"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "subscription_date": {
            "type": "string",
            "format": "date",
            "title": "Subscription Date"
          },
          "effective_date": {
            "type": "string",
            "format": "date",
            "title": "Effective Date"
          },
          "annual_premium": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Annual Premium"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "guarantee": {
            "type": "string",
            "title": "Guarantee"
          },
          "segment": {
            "type": "string",
            "title": "Segment"
          },
          "file_rate_year_1": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Rate Year 1"
          }
        },
        "type": "object",
        "required": [
          "contract_id",
          "product",
          "state",
          "subscription_date",
          "effective_date",
          "annual_premium",
          "end_date",
          "guarantee",
          "segment",
          "file_rate_year_1"
        ],
        "title": "ContractRecordOut"
      },
      "ExclusionOut": {
        "properties": {
          "contract_id": {
            "type": "string",
            "title": "Contract Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "contract_id",
          "code",
          "label",
          "detail"
        ],
        "title": "ExclusionOut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LinearPolicyOut": {
        "properties": {
          "excluded_movement_prefixes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Excluded Movement Prefixes"
          },
          "excluded_premium_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Excluded Premium Types"
          },
          "excluded_products": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Excluded Products"
          },
          "first_year_filter": {
            "type": "boolean",
            "title": "First Year Filter"
          }
        },
        "type": "object",
        "required": [
          "excluded_movement_prefixes",
          "excluded_premium_types",
          "excluded_products",
          "first_year_filter"
        ],
        "title": "LinearPolicyOut"
      },
      "PerimeterDetailsOut": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "insurer": {
            "type": "string",
            "title": "Insurer"
          },
          "line_of_business": {
            "type": "string",
            "title": "Line Of Business"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "key_strategy": {
            "type": "string",
            "title": "Key Strategy"
          },
          "key_strategy_label": {
            "type": "string",
            "title": "Key Strategy Label"
          },
          "exposure_base": {
            "type": "string",
            "title": "Exposure Base"
          },
          "exposure_base_label": {
            "type": "string",
            "title": "Exposure Base Label"
          },
          "rate_source": {
            "type": "string",
            "title": "Rate Source"
          },
          "rate_source_label": {
            "type": "string",
            "title": "Rate Source Label"
          },
          "tariff_regularisation": {
            "type": "boolean",
            "title": "Tariff Regularisation"
          },
          "seniority_months": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seniority Months"
          },
          "old_contract_threshold_years": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Old Contract Threshold Years"
          },
          "max_future_effect_months": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Future Effect Months"
          },
          "rate_overrides": {
            "items": {
              "$ref": "#/components/schemas/RateOverrideOut"
            },
            "type": "array",
            "title": "Rate Overrides"
          },
          "product_rules": {
            "items": {
              "$ref": "#/components/schemas/ProductRuleOut"
            },
            "type": "array",
            "title": "Product Rules"
          },
          "linear": {
            "$ref": "#/components/schemas/LinearPolicyOut"
          },
          "products": {
            "items": {
              "$ref": "#/components/schemas/ProductOut"
            },
            "type": "array",
            "title": "Products"
          }
        },
        "type": "object",
        "required": [
          "code",
          "label",
          "description",
          "insurer",
          "line_of_business",
          "active",
          "key_strategy",
          "key_strategy_label",
          "exposure_base",
          "exposure_base_label",
          "rate_source",
          "rate_source_label",
          "tariff_regularisation",
          "seniority_months",
          "old_contract_threshold_years",
          "max_future_effect_months",
          "rate_overrides",
          "product_rules",
          "linear",
          "products"
        ],
        "title": "PerimeterDetailsOut"
      },
...
      "PreviousMonthInput": {
        "properties": {
          "state": {
            "type": "string",
            "maxLength": 40,
            "minLength": 1,
            "title": "State"
          },
          "annual_premium": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 10000000.0,
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Annual Premium"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "state"
        ],
        "title": "PreviousMonthInput",
        "description": "Situation du même contrat le mois précédent : seuls l'état, la prime et la date de fin peuvent différer."
      },
      "ProductOut": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "year_1": {
            "items": {
              "$ref": "#/components/schemas/RateWindowOut"
            },
            "type": "array",
            "title": "Year 1"
          },
          "year_2_plus": {
            "items": {
              "$ref": "#/components/schemas/RateWindowOut"
            },
            "type": "array",
            "title": "Year 2 Plus"
          }
        },
        "type": "object",
        "required": [
          "code",
          "label",
          "year_1",
          "year_2_plus"
        ],
        "title": "ProductOut"
      },
      "ProductRuleOut": {
        "properties": {
          "contract_prefix": {
            "type": "string",
            "title": "Contract Prefix"
          },
          "product": {
            "type": "string",
            "title": "Product"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "contract_prefix",
          "product",
          "reason"
        ],
        "title": "ProductRuleOut"
      },
      "RateOverrideOut": {
        "properties": {
          "rate": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Rate"
          },
          "display": {
            "type": "string",
            "title": "Display"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "product": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product"
          },
          "guarantee": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Guarantee"
          },
          "segment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Segment"
          }
        },
        "type": "object",
        "required": [
          "rate",
          "display",
          "reason",
          "product",
          "guarantee",
          "segment"
        ],
        "title": "RateOverrideOut"
      },
...
      "SampleContractOut": {
        "properties": {
          "perimeter": {
            "type": "string",
            "title": "Perimeter"
          },
          "month": {
            "type": "string",
            "title": "Month"
          },
          "previous_month": {
            "type": "string",
            "title": "Previous Month"
          },
          "contract_id": {
            "type": "string",
            "title": "Contract Id"
          },
          "previous_records": {
            "items": {
              "$ref": "#/components/schemas/ContractRecordOut"
            },
            "type": "array",
            "title": "Previous Records"
          },
          "current_records": {
            "items": {
              "$ref": "#/components/schemas/ContractRecordOut"
            },
            "type": "array",
            "title": "Current Records"
          },
          "premiums": {
            "items": {
              "$ref": "#/components/schemas/PremiumRecordOut"
            },
            "type": "array",
            "title": "Premiums"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/CommissionLineOut"
            },
            "type": "array",
            "title": "Lines"
          },
          "exclusions": {
            "items": {
              "$ref": "#/components/schemas/ExclusionOut"
            },
            "type": "array",
            "title": "Exclusions"
          }
        },
        "type": "object",
        "required": [
          "perimeter",
          "month",
          "previous_month",
          "contract_id",
          "previous_records",
          "current_records",
          "premiums",
          "lines",
          "exclusions"
        ],
        "title": "SampleContractOut"
      },
      "SimulationRequest": {
        "properties": {
          "perimeter": {
            "type": "string",
            "maxLength": 40,
            "minLength": 1,
            "title": "Perimeter"
          },
          "month": {
            "type": "string",
            "pattern": "^\\d{4}-(0[1-9]|1[0-2])$",
            "title": "Month",
            "description": "Mois de calcul, AAAA-MM"
          },
          "contract": {
            "$ref": "#/components/schemas/ContractInput"
          },
          "previous_month": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PreviousMonthInput"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "perimeter",
          "month",
          "contract"
        ],
        "title": "SimulationRequest"
      },
...
      "TotalsOut": {
        "properties": {
          "precompte": {
            "$ref": "#/components/schemas/AmountOut"
          },
          "reprises": {
            "$ref": "#/components/schemas/AmountOut"
          },
          "lineaire": {
            "$ref": "#/components/schemas/AmountOut"
          },
          "net": {
            "$ref": "#/components/schemas/AmountOut"
          }
        },
        "type": "object",
        "required": [
          "precompte",
          "reprises",
          "lineaire",
          "net"
        ],
        "title": "TotalsOut"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}